Implemented Dot language parsing
